added NginxConfigFileGenerator to pass in custom server.conf templates to NginxController
