+++ software/webapp/src/main/java/brooklyn/entity/proxy/nginx/NginxController.java
@@ -37,6 +37,7 @@ import brooklyn.event.basic.Sensors;
 import brooklyn.util.flags.SetFromFlag;
 
 import com.google.common.collect.ImmutableMap;
+import com.google.common.reflect.TypeToken;
 
 /**
  * An entity that represents an Nginx proxy (e.g. for routing requests to servers in a cluster).
@@ -104,9 +105,12 @@ public interface NginxController extends AbstractController, HasShortName {
     ConfigKey<String> WITH_CC_OPT = ConfigKeys.newStringConfigKey(
             "nginx.install.withCcOpt", "String to pass in with --with-cc-opt=\"<val>\"", "-I /usr/local/include");
 
+    @SetFromFlag("configGenerator")
+    ConfigKey<NginxConfigFileGenerator> SERVER_CONF_GENERATOR = ConfigKeys.newConfigKey(NginxConfigFileGenerator.class,
+            "nginx.config.generator", "The server.conf generator class", new NginxDefaultConfigGenerator());
+
     @SetFromFlag("configTemplate")
-    ConfigKey<String> SERVER_CONF_TEMPLATE_URL = ConfigKeys.newStringConfigKey(
-            "nginx.config.templateUrl", "The server.conf configuration file URL (FreeMarker template)");
+    ConfigKey<String> SERVER_CONF_TEMPLATE_URL = NginxTemplateConfigGenerator.SERVER_CONF_TEMPLATE_URL;
 
     @SetFromFlag("staticContentArchive")
     ConfigKey<String> STATIC_CONTENT_ARCHIVE_URL = ConfigKeys.newStringConfigKey(
